+++ common.h
@@ -12,7 +12,7 @@
 #define MIN_GTK_RUNTIME(major, minor, micro) (!gtk_check_version(major, minor, micro))
 
 #define APPNAME "pingpath"
-#define VERSION "1.0.10"
+#define VERSION "1.0.11"
 #define APPVER  APPNAME "-" VERSION
 
 extern locale_t locale, localeC;