+++ tabs/aux.h
@@ -3,9 +3,10 @@
 
 #include "common.h"
 
-typedef struct tab_widget { GtkWidget *w; const char *css, *col; } t_tab_widget;
-#define TAB_ELEM_WITH(tabw, maker, cssclass, color, retval) { (tabw).css = cssclass; (tabw).col = color; \
-  (tabw).w = maker; g_return_val_if_fail((tabw).w, retval); }
+typedef struct tab_widget {
+  GtkWidget *w;
+  const char *css, *col;
+} t_tab_widget;
 
 typedef struct t_sa_desc {
   GSimpleAction* sa;
@@ -26,8 +27,8 @@ typedef struct tab {
   GActionEntry act[POP_MENU_NDX_MAX];
 } t_tab;
 
-gboolean basetab_init(t_tab *tab, GtkWidget* (*make_dyn)(void), GtkWidget* (*make_extra)(void));
-gboolean drawtab_init(t_tab *tab, const char *color, GSList *layers, guint ndx);
+gboolean basetab_init(t_tab *tab, GtkWidget* (*make_dyn)(void), GtkWidget* (*make_extra)(void)); // NONNULL(1, 2)
+gboolean drawtab_init(t_tab *tab, const char *color, GSList *layers, uint ndx); // NONNULL(1, 3)
 void tab_setup(t_tab *tab);
 void tab_color(t_tab *tab);
 void tab_reload_theme(void);